Add url converter interface for easier testing; move converter's package
